task_canceler: remove assert
This assert was all nice, but it triggers spuriously.

The task canceler gives two properties:

- No more asynchronous code will run on AsyncCore when it is dead, in
  particular all async lambdas that capture `this`
- No more event handler will run after AsyncCore is destroyed. This
  ensures that after a call to tanker_destroy, no more event handler
  will run, even if the event occurred *before* the destruction. This is
  very useful to write code that is not segfaulty

However, this assert triggers in the following case: tanker_destroy is
called from an event handler. The task canceler will cancel the event
handler that is currently running which is a no-op. It is not harmful
though as the event handler is not called synchronously by AsyncCore, so
it is fine to destroy it.