Fix linker version scripts not actually hiding any symbols!
Declaring symbols as STB_GLOBAL is actually the default behavior.
That is, if we only tell the linker which symbols we want to keep,
it dutifully does that and nothing more, in pure bad UX tradition.

This fix was made necessary by a symbol conflict in the Ruby binding.
The native SDK links with SSL symbols internally, and those symbols
could be interposed by libcrypto.so.1.1 in all sorts of unfortunate
ways.
Note that we could used fancier symbol visibilities instead of local
binding to fix the issue for those symbols, but since this was
already the intent of the version script, we might as well fix it here