Revocation: pull user before nuking and throwing
Before this commit, we used to nuke the storage and throw just after the
server returned a "device_revoked" error. The correct behavior would be
to pull the user's blocks and verify the revocation before taking any
action.

This commit changes how AsyncCore reacts to the error. There are two
different errors that it can receive:

- Errc::DeviceRevoked which is thrown by the User Updater after a
  revocation of the current device has been *verified*
- AppdErrc::DeviceRevoked which is thrown by HttpClient and corresponds
  only to the server telling us that the operation was rejected

When AsyncCore catches AppdErrc::DeviceRevoked, it triggers a user
verification that will in turn throw Errc::DeviceRevoked. It will nuke
the device and rethrow that last exception.

Also, authenticating with a revoked device is not an error (AWARDINAE,
like C++), you just get a flag in the response. The token you receive
though is only usable with the revoked-device-history API, you can't
get your history the normal way, it will throw a 401 error.