AsyncCore: start the log threadpool lazily
Ruby on Rails starts the interpreter and forks it. The problem is that
we used to start the log thread upon call to set_log_handler, and for
some reason, that thread isn't forked, or maybe it is but it crashes, or
I don't know.

So to solve this, we decided to start the thread later. This means that
when the Tanker Ruby SDK is loaded, we dlopen the lib, but don't start
any work, to be fork-safe.