Add UNSUPPORTED_GROUP_VERSION error
The GroupUpdate block, which allows removing users from groups,
is not compatible with the old Version 1 group block format.
This error signals that a group contains V1 blocks,
and the group update feature is unsupported as a result.

To resolve this error, create a new group with the same members.