Teach Group Accessor to compute an internal group's member list
This is required to create GroupUpdate blocks.

Previously, we only had GroupCreation and GroupAddition, which don't
need to know the current members, because they only add new ones.
In contrast, making a GroupUpdate block means rotating the group keys
and removing existing users, so we must get the current members first.