Add types for members extracted from a Group V3 block
When doing a group update, we have to carry over existing users from
the existing group, add new users, and then encrypt them the new key.

This is not the same as the full UserGroup*Member structures used in
the actual blocks, only the bare (provisional/user) identity from it.