Only push V3 group blocks if the group is already a V3 group
The GroupUpdate block is not currently supported, note however
the only way that a group that previously contained provisional users
can be used with the Group Update feature is if it was created with a
GroupCreation V3 block, or if it was reset with a GroupUpdate V3.

For example, a V2 group can then have a GroupAddition V3, but if it
still contains V2 provisional users it will still need a migration.

This means that this commit is *NOT* enough to transparently support
the upgrade to V3 blocks, except for new groups created in V3.