cstream: call stream callback without async
We used to call the callback through tc::async because Java doesn't like
to be called from a coroutine stack. tc::async posted an asynchronous
work to get back to the thread stack.

While we need to change stack, we don't need this call to be
asynchronous and have the asio queue and future overhead.
dispatch_on_thread_context solves the problem by calling the lambda on
the thread stack, without adding asynchronicity.