bench: Switch MR bench job to manual to not re-bench on every push
The nightly bench/master job is STILL auto, this is just the optional
bench job for WIP sdk-native merge requests.

Ideally, we'd want the job to trigger automatically when a MR is
opened, and then leave it on manual for further pushes (since most
of the pushes will usually be small review fixes).

But I can't find a way to do that, so for now avoid overloading
the runners and spamming messages by keeping it manual.