UserAccessor: fix request of provisional users
When requesting provisional identities, we would ask them all twice.
Then we have a loop that iterates over two lists at the same time, one
has duplicates, the other doesn't.

So if we request A and B, one list contains A, A, B, B, and the other
just A, B. So we match A with A, and B with A too, and we discard the B,
B portion altogether.