fix(v6): added the version number to additionalData
Adding the version number to additionalData aims at changing the MAC
so that its decryption will fail if one tries to alter the version byte
in the encrypted data.

Note: in `decrypt`, I use `encryptedData[0]` instead of
`EncryptorV6::version()` in order to check if the version byte was
altered.
Since this encryptor has a minimal overhead, this should not cause a segfault.