AsyncCore: fix race between stop and other operations
This concurrency is handled by tconcurrent after the future is added to
the task canceler. However, if stop() and an operation are called at the
same time on different thread, the operation may not be canceled.

The previous code with the _stopping variable was broken for different
reasons:

- The check of _stopping and the addition to the task_canceler must be
  atomic (this fixes the bug described above)
- I forgot a return statement in the if
- make_execptional_future takes an exception, not an exception_ptr

I tried to write a test for this with 3 threads calling encrypt() in a
loop, and one thread calling stop() and start() in a loop. I couldn't
reproduce the issue, so I didn't commit the test. For reference, this
crash came from a concurrency test I wrote for the Ruby SDK.