test(setVerificationMethod): E2E passphrase tests
E2E verification methods are special, because we cannot have them
at the same time as non-E2E methods. When switching, the old methods
are removed and should stop working.

There are 4 possible cases for setVerificationMethod (switching or not,
using an E2E method or not). This commit is where we test all of that.