fix: Trigger bridge jobs on matching branch to fix release tags
Previously, we would detect the upstream branch and try to switch
the branch in the downstream in a run-ci script.
Doing it in the upstream is cleaner, as all the variables for the
downstream pipeline will be actually set to the right branch.

This also fixes tags for downstreams always being set on the master
branch, instead of the matching release branch (when it exists)