fix(.gitlab-ci): Use prod artifactory for release/ branches
Today only release pipelines on master use the prod artifactory,
this doesn't match the release process documentation that
optionally allows using a release branch.
We want to use the same artifactory repo, whether the pipeline
ran on master or on a release/ branch.

For alphas and betas, it's OK to have them also be on the
prod artifactory, we already explicitly check for alpha and
beta versions in the deploy script, so only stable releases can
become the conan latest-stable reference.

It is still possible to use artifactory staging by starting a
pipeline on a non release branch, e.g. feat/mybranch