fix(.gitlab-ci): Split deploy/tag from deploy job, use deploy/tag rule
During release pipelines the deploy job must always run,
because the downstream deploy jobs have a `needs` dependency on it.

However, it is possible to run release pipelines with no
$SDK_NATIVE_RELEASE_VERSION set, but where we still want to release
downstream bindings.
In this case we must not try to tell gitlab to make a native release,
that is how the "v" tags were created.

However, there is no way to make a `needs` conditional on a variable
not being empty. This is sad, so instead we let the deploy job
always run (to satisfy needs), but separate out the deploy/tag job
that actually creates the gitlab release, and we add the rule
.rules/deploy/native/tag to skip if SDK_NATIVE_RELEASE_VERSION is null