feat: SimpleResourceId replaces ResourceId, which is now a variant
ResourceId is still the value returned by getResourcedId(),
which can be either a SimpleResourceId or a CompositeResourceId.

Places in the code that work with a single key at a time
(so, everything excluding getResourceId() and share() for now)
can continue to work with SimpleResourceId directly,
this includes existing encryption formats and accessors.