feat: CompositeResourceId and SimpleResourceId types
With transparent sessions, the resource ID can now be a composite
of the individual resource's ID (used to share just the resource)
and the session ID, and thus can refer to multiple keys.

The simple resource ID is the same as the old resource ID type,
and only identifies to a single (resource or session) key.