feat: Introduce HttpHeaderMap class for C++ requests
This is technically a multimap for compat with the HTTP spec,
but in practice we almost always want to get/set a single header,
so there are user-friendly convenience functions.

Some minimal iterator interface is exposed for the more complicated
cases, to show that it's possible, but we probably won't need it.