feat: BREAKING the ctanker API to take and return list of headers
Note that the lifetime of a HTTP request/response object is
entirely tied to the C++ side asynchronously waiting for the response.
So we don't have to specify any C function for allocating or freeing
C HTTP header objects.

The C++ side just keeps a C++ object alive, and gives a pointer to it.
Normal C++ destructors clean it up after it gets the response.

The FFI side does the same. It allocates however it wants,
and it can free its headers and other C strings after responding.